Move by classes to their own package Refactor by button and by tabset to not use ID's
