+++ wcomponents-test-lib/src/main/java/com/github/bordertech/wcomponents/test/selenium/by/ByButtonImage.java
@@ -1,0 +1,17 @@
+package com.github.bordertech.wcomponents.test.selenium.by;
+
+import java.util.List;
+import org.openqa.selenium.By;
+import org.openqa.selenium.SearchContext;
+import org.openqa.selenium.WebElement;
+
+/**
+ * Finds a button by the image on the button.
+ */
+public class ByButtonImage extends By {
+
+	@Override
+	public List<WebElement> findElements(final SearchContext context) {
+		return null;
+	}
+}
