+++ wcomponents-test-lib/src/main/java/com/github/bordertech/wcomponents/test/selenium/by/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains classes that help test WComponent applications.
+ */
+package com.github.bordertech.wcomponents.test.selenium.by;
