+++ wcomponents-test-lib/src/test/java/com/github/bordertech/wcomponents/test/WPhoneNumberFieldWebElementTest.java
@@ -1,8 +1,8 @@
 package com.github.bordertech.wcomponents.test;
 
 import com.github.bordertech.wcomponents.test.components.WPhoneNumberFieldUI;
-import com.github.bordertech.wcomponents.test.selenium.ByButtonText;
-import com.github.bordertech.wcomponents.test.selenium.ByLabel;
+import com.github.bordertech.wcomponents.test.selenium.by.ByButtonText;
+import com.github.bordertech.wcomponents.test.selenium.by.ByLabel;
 import com.github.bordertech.wcomponents.test.selenium.MultiBrowserRunner;
 import com.github.bordertech.wcomponents.test.selenium.SeleniumJettyTestCase;
 import com.github.bordertech.wcomponents.test.selenium.driver.SeleniumWComponentsWebDriver;
