+++ wcomponents-test-lib/src/test/java/com/github/bordertech/wcomponents/test/WTabSetWebElementTest.java
@@ -1,7 +1,7 @@
 package com.github.bordertech.wcomponents.test;
 
 import com.github.bordertech.wcomponents.test.components.WTabSetUI;
-import com.github.bordertech.wcomponents.test.selenium.ByWTabSet;
+import com.github.bordertech.wcomponents.test.selenium.by.ByWTabSet;
 import com.github.bordertech.wcomponents.test.selenium.MultiBrowserRunner;
 import com.github.bordertech.wcomponents.test.selenium.SeleniumJettyTestCase;
 import com.github.bordertech.wcomponents.test.selenium.driver.SeleniumWComponentsWebDriver;
